+++ .deploy_cfg
@@ -1,2 +1,2 @@
 [auth]
-hostname = 169.254.234.126
+hostname = 10.0.74.95
