+++ robot.py
@@ -1,25 +1,36 @@
 #!/usr/bin/env python3
+
 import wpilib
+from wpilib import RobotDrive
+from wpilib import SPI
 from xbox import XboxController
 from wpilib.smartdashboard import SmartDashboard
-from components.drive import driveTrain
+#from components.drive import driveTrain
+from ctre.cantalon import CANTalon
 from robotpy_ext.autonomous.selector import AutonomousModeSelector
-from wpilib import USBCamera, CameraServer
+
+## from wpilib import USBCamera, CameraServer
+
 
 CONTROL_LOOP_WAIT_TIME = 0.025
 
 class MyRobot(wpilib.SampleRobot):
 
     def robotInit(self):
+
         self.controller = XboxController(0)
-        self.controller2 = XboxController(1)
+        #self.controller2 = XboxController(1)
+        self.gyro = wpilib.ADXRS450_Gyro(0)
 
-        #self.lmotor = wpilib.CANTalon(1)
-        #self.rmotor = wpilib.CANTalon(0)
+        self.lfmotor = CANTalon(1)
+        self.lbmotor = CANTalon(2)
+        self.rfmotor = CANTalon(3)
+        self.rbmotor = CANTalon(0)
 
-        self.drive = driveTrain(self)
+        self.drive = wpilib.RobotDrive(self.lfmotor, self.lbmotor, self.rfmotor, self.rbmotor)
+        #self.drive = RobotDrive()
 
-        self.drive.reset()
+        #self.drive.reset()
 
         self.dashTimer = wpilib.Timer()     # Timer for SmartDashboard updating
         self.dashTimer.start()
@@ -33,69 +44,60 @@ class MyRobot(wpilib.SampleRobot):
         self.dash = SmartDashboard()
         self.autonomous_modes = AutonomousModeSelector('autonomous', self.components)
 
-        self.drive.log()
+        #self.drive.log()
 
 
     def disabled(self):
-        self.drive.reset()
+
+        #self.drive.reset()
         #self.drive.disablePIDs()
 
         while self.isDisabled():
             wpilib.Timer.delay(0.01)              # Wait for 0.01 seconds
 
     def autonomous(self):
-        self.drive.reset()
-        self.drive.enablePIDs()
-
-        while self.isAutonomous() and self.isEnabled():
-
-            # Run the actual autonomous mode
-            self.drive.log()
-            self.autonomous_modes.run()
-
-    def operatorControl(self):
-        # Resetting encoders
 
         self.drive.reset()
         #self.drive.enablePIDs()
 
+        while self.isAutonomous() and self.isEnabled():
+
+            # Run the actual autonomous mode
+            #self.drive.log()
+            self.autonomous_modes.run()
+
+    def operatorControl(self):
+        # Resetting encoders
+        wpilib.LiveWindow.run()
+        #self.drive.enablePIDs()
+        self.gyro.reset()
+        self.gyro.calibrate()
+
         while self.isOperatorControl() and self.isEnabled():
-            self.drive.xboxTankDrive(self.controller.getLeftY(), self.controller.getRightY(), self.controller.getLeftBumper(), self.controller.getRightBumper(), self.controller.getLeftTrigger(), self.controller.getRightTrigger())
+            self.drive.mecanumDrive_Cartesian(self.controller.getLeftX(), self.controller.getLeftY(), self.controller.getRightX(), 0)
+            wpilib.SmartDashboard.putNumber("GyroAngle",self.gyro.getAngle())
+            #wpilib.SmartDashboard.putNumber("getAccumulatorValue",self.gyro.spi.getAccumulatorValue())
+            #wpilib.SmartDashboard.putNumber("kDegreePerSecond",self.gyro.kDegreePerSecondPerLSB)
+            #wpilib.SmartDashboard.putNumber("kSamplePeriod",self.gyro.kSamplePeriod)
+            #wpilib.SmartDashboard.putNumber("GyroRate", self.gyro.getRate())
 
-            self.drive.log()
-
-            wpilib.Timer.delay(CONTROL_LOOP_WAIT_TIME)
+            '''
+            if(self.gyro.spi == None):
+                wpilib.SmartDashboard.putNumber("SPIValue", 0)
+            else:
+                wpilib.SmartDashboard.putNumber("SPIValue", 1)
+            '''
 
     def test(self):
+
         wpilib.LiveWindow.run()
 
         self.drive.reset()
-        self.drive.enablePIDs()
+        #self.drive.enablePIDs()
 
         while self.isTest() and self.isEnabled():
 
-            self.drive.xboxTankDrive(self.controller.getLeftY(), self.controller.getRightY(), self.controller.getLeftBumper(), self.controller.getRightBumper(), self.controller.getLeftTrigger(), self.controller.getRightTrigger())
-
-    '''
-    def checkPixy():
-        distanceFromCenter = 0
-        closestBallArea = 0 #meaningless number that any result returned by the function can always beat
-        biggestBallID = None
-        blocks = self.pixy.getBlocks()
-
-        for i in range(0, len(blocks)):
-            area = blocks[i].getArea()
-            if(area > closestBallArea):
-                closestBallArea = area
-                biggestBallID = i
-       if(controller.getLeftTriggerRaw() > 0.75):
-            distanceFromCenter = blocks[BiggestballID] - 180
-
-            if(distanceFromCenter < 0):#turn right
-                self.drive.turnAngle(-2)
-            elif(distanceFromCenter > 0):#turn left
-                self.drive.turnAngle(2)
-    '''
+            self.drive.xboxTankDrive(self.controller.getLeftY(), self.controller.getRightY())
 
 if __name__ == "__main__":
     wpilib.run(MyRobot)
